refactor(ports): each port's main file moves outside its own directory
ports/x/x.rb was the redundant shape — the port's face buried inside a
same-named folder. Each face now sits at ports/x.rb with its .port
declaration beside it, and only persistence keeps a directory (its five
collaborators). Query's three already-distinct modules become three
files: query.rb (the boundary), query/ordering.rb, query/in_memory.rb.
A new ports.rb wrapper gives lib/hecksagain.rb one require for the
whole layer. The Folder adapter's library glob follows the .port files
up a level.
