+++ bin/ir_rust
@@ -61,8 +61,8 @@ path = ARGV[0] or abort "usage: bin/ir_rust <bluebook>"
 #                                   is checked against come from ONE reading.
 ROOT = File.expand_path("..", __dir__)
 PORTS = %w[
-  lib/hecksagain/ports/persistence/persistence.port
-  lib/hecksagain/ports/extraction/extraction.port
+  lib/hecksagain/ports/persistence.port
+  lib/hecksagain/ports/extraction.port
   lib/hecksagain/adapters/driven/memory/memory.adapter
   lib/hecksagain/adapters/driven/prism/prism.adapter
 ].freeze
