+++ lib/hecksagain.rb
@@ -9,11 +9,7 @@ require_relative "hecksagain/query_specification/common/null_policy"
 require_relative "hecksagain/query_specification/common/dsl"
 require_relative "hecksagain/query_specification/read_model/specification"
 
-require_relative "hecksagain/ports/loading/loading"
-require_relative "hecksagain/ports/persistence/persistence"
-require_relative "hecksagain/ports/query/query"
-require_relative "hecksagain/ports/projection/projection"
-require_relative "hecksagain/ports/extraction/extraction"
+require_relative "hecksagain/ports"
 require_relative "hecksagain/bluebook/expression/canonical_form"
 require_relative "hecksagain/bluebook/expression/resolver"
 require_relative "hecksagain/bluebook/expression/evaluator"
