+++ lib/hecksagain/bluebook/meta_validator.rb
@@ -177,8 +177,8 @@ module Hecksagain
       def self.load_grammar_into(registry)
         @bootstrapping = true
         Hecksagain.with_registry(registry) do
-          Kernel.load(File.expand_path("../ports/persistence/persistence.port", __dir__))
-          Kernel.load(File.expand_path("../ports/extraction/extraction.port", __dir__))
+          Kernel.load(File.expand_path("../ports/persistence.port", __dir__))
+          Kernel.load(File.expand_path("../ports/extraction.port", __dir__))
           Kernel.load(File.expand_path("../adapters/driven/memory/memory.adapter", __dir__))
           Kernel.load(File.expand_path("../adapters/driven/prism/prism.adapter", __dir__))
           GRAMMAR_FILES.each { |file| Kernel.load(file) }
