+++ lib/hecksagain/ports/extraction.rb
