+++ lib/hecksagain/ports/projection.rb
