+++ spec/spec_helper.rb
@@ -14,8 +14,8 @@ require "hecksagain"
 module InMemoryDomain
   ROOT             = File.expand_path("..", __dir__)
   PIZZAS_BLUEBOOK  = File.join(ROOT, "examples/pizzas/bluebook/pizzas.bluebook")
-  PERSISTENCE_PORT = File.join(ROOT, "lib/hecksagain/ports/persistence/persistence.port")
-  EXTRACTION_PORT  = File.join(ROOT, "lib/hecksagain/ports/extraction/extraction.port")
+  PERSISTENCE_PORT = File.join(ROOT, "lib/hecksagain/ports/persistence.port")
+  EXTRACTION_PORT  = File.join(ROOT, "lib/hecksagain/ports/extraction.port")
   MEMORY_ADAPTER   = File.join(ROOT, "lib/hecksagain/adapters/driven/memory/memory.adapter")
   PRISM_ADAPTER    = File.join(ROOT, "lib/hecksagain/adapters/driven/prism/prism.adapter")
